[SCB-2029] tiny improve to make extend filter features easier

+++ core/src/main/java/org/apache/servicecomb/core/exception/converter/ConstraintViolationExceptionConverter.java
@@ -35,7 +35,7 @@ import com.netflix.config.DynamicPropertyFactory;
 import com.netflix.config.DynamicStringProperty;
 
 public class ConstraintViolationExceptionConverter implements ExceptionConverter<ConstraintViolationException> {
-  public static final short ORDER = Short.MAX_VALUE;
+  public static final int ORDER = Short.MAX_VALUE;
 
   public static final String KEY_CODE = "servicecomb.filters.validate.code";
