+++ core/src/main/java/org/apache/servicecomb/core/exception/converter/InvocationExceptionConverter.java
@@ -24,7 +24,7 @@ import org.apache.servicecomb.core.exception.ExceptionConverter;
 import org.apache.servicecomb.swagger.invocation.exception.InvocationException;
 
 public class InvocationExceptionConverter implements ExceptionConverter<InvocationException> {
-  public static final byte ORDER = Byte.MAX_VALUE;
+  public static final int ORDER = Byte.MAX_VALUE;
 
   @Override
   public int getOrder() {
